Refactoring create project processing
