+++ app/cli/createproject/createproject_cli_workflow.go
@@ -1,16 +1,21 @@
 package createproject
 
 import (
+	"k8s-management-go/app/actions/createproject"
+	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/loggingstate"
 )
 
 func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
+	var projectConfig models.ProjectConfig
+	projectConfig.CreateDeploymentOnlyProject = deploymentOnly
+
 	// Start project wizard
 	loggingstate.AddInfoEntry("Starting Project Wizard: Dialogs...")
 
 	// Ask for namespace
 	loggingstate.AddInfoEntry("-> Ask for namespace...")
-	namespace, err := NamespaceWorkflow()
+	projectConfig.Namespace, err = NamespaceWorkflow()
 	if err != nil {
 		return err
 	}
@@ -18,23 +23,17 @@ func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
 
 	// Ask for IP address
 	loggingstate.AddInfoEntry("-> Ask for IP address...")
-	ipAddress, err := IpAddressWorkflow()
+	projectConfig.IpAddress, err = IpAddressWorkflow()
 	if err != nil {
 		return err
 	}
 	loggingstate.AddInfoEntry("-> Ask for IP address...done")
 
-	// declare vars for next if statement
-	var jenkinsSystemMsg string
-	var jobsCfgRepo string
-	var selectedCloudTemplates []string
-	var existingPvc string
-
 	// if it is not only a deployment project ask for other Jenkins related vars
-	if !deploymentOnly {
+	if !projectConfig.CreateDeploymentOnlyProject {
 		// Select cloud templates
 		loggingstate.AddInfoEntry("-> Ask for cloud templates...")
-		selectedCloudTemplates, err = CloudTemplatesWorkflow()
+		projectConfig.SelectedCloudTemplates, err = CloudTemplatesWorkflow()
 		if err != nil {
 			return err
 		}
@@ -42,7 +41,7 @@ func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
 
 		// Ask for existing persistent volume claim (PVC)
 		loggingstate.AddInfoEntry("-> Ask for persistent volume claim...")
-		existingPvc, err = PersistentVolumeClaimWorkflow()
+		projectConfig.ExistingPvc, err = PersistentVolumeClaimWorkflow()
 		if err != nil {
 			return err
 		}
@@ -50,7 +49,7 @@ func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
 
 		// Ask for Jenkins system message
 		loggingstate.AddInfoEntry("-> Ask for Jenkins system message...")
-		jenkinsSystemMsg, err = JenkinsSystemMessageWorkflow(namespace)
+		projectConfig.JenkinsSystemMsg, err = JenkinsSystemMessageWorkflow(projectConfig.Namespace)
 		if err != nil {
 			return err
 		}
@@ -58,7 +57,7 @@ func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
 
 		// Ask for Jobs Configuration repository
 		loggingstate.AddInfoEntry("-> Ask for jobs configuration repository...")
-		jobsCfgRepo, err = JenkinsJobsConfigRepositoryWorkflow()
+		projectConfig.JobsCfgRepo, err = JenkinsJobsConfigRepositoryWorkflow()
 		if err != nil {
 			return err
 		}
@@ -68,7 +67,7 @@ func ProjectWizardWorkflow(deploymentOnly bool) (err error) {
 
 	// Process data and create project
 	loggingstate.AddInfoEntry("Starting Project Wizard: Template processing...")
-	err = ProcessProjectCreate(namespace, ipAddress, jenkinsSystemMsg, jobsCfgRepo, existingPvc, selectedCloudTemplates, deploymentOnly)
+	err = createproject.ActionProcessProjectCreate(projectConfig)
 	if err != nil {
 		loggingstate.AddInfoEntry("Starting Project Wizard: Template processing...failed")
 	}
