+++ app/models/project_config.go
@@ -1,0 +1,11 @@
+package models
+
+type ProjectConfig struct {
+	Namespace                   string
+	IpAddress                   string
+	JenkinsSystemMsg            string
+	JobsCfgRepo                 string
+	ExistingPvc                 string
+	SelectedCloudTemplates      []string
+	CreateDeploymentOnlyProject bool
+}
